Move new tests to separate module
